+++ game_patch/multi/bagman.cpp
@@ -525,7 +525,7 @@ void bagman_do_frame()
             if (rf::gameseq_get_state() == rf::GameState::GS_GAMEPLAY
                 && g_bagman_info.score_tick.elapsed()) {
                 rf::player_add_score(g_bagman_info.carrier, 1);
-                if (gt_is_tbm()) {
+                if (gt_is_tbag()) {
                     if (g_bagman_info.carrier->team == 0) {
                         g_bagman_info.red_team_score++;
                     } else {