+++ game_patch/multi/gametype.cpp
@@ -30,10 +30,10 @@ static char run_name[] = "RUN";
 static char* run_slot = run_name;
 static char esc_name[] = "ESC";
 static char* esc_slot = esc_name;
-static char bm_name[] = "BM";
-static char* bm_slot = bm_name;
-static char tbm_name[] = "TBM";
-static char* tbm_slot = tbm_name;
+static char bag_name[] = "BAG";
+static char* bag_slot = bag_name;
+static char tbag_name[] = "TBAG";
+static char* tbag_slot = tbag_name;
 // UNK is the sentinel; new game types must be added above
 static char unk_name[] = "UNK";
 static char* unk_slot = unk_name;
@@ -73,8 +73,8 @@ void populate_gametype_table() {
     g_af_gametype_names[rf::NG_TYPE_REV]    = &rev_slot;
     g_af_gametype_names[rf::NG_TYPE_RUN]    = &run_slot;
     g_af_gametype_names[rf::NG_TYPE_ESC]    = &esc_slot;
-    g_af_gametype_names[rf::NG_TYPE_BM]     = &bm_slot;
-    g_af_gametype_names[rf::NG_TYPE_TBM]    = &tbm_slot;
+    g_af_gametype_names[rf::NG_TYPE_BAG]     = &bag_slot;
+    g_af_gametype_names[rf::NG_TYPE_TBAG]    = &tbag_slot;
     g_af_gametype_names[rf::NG_TYPE_UNK]    = &unk_slot;
 
     for (int i = 0; i < 5; ++i) {
@@ -122,9 +122,9 @@ bool multi_game_type_is_team_type(rf::NetGameType game_type)
         case rf::NG_TYPE_DC:
         case rf::NG_TYPE_REV:
         case rf::NG_TYPE_ESC:
-        case rf::NG_TYPE_TBM:
+        case rf::NG_TYPE_TBAG:
             return true;
-        default: // DM, RUN, BM
+        default: // DM, RUN, BAG
             return false;
     }
 }
@@ -216,19 +216,19 @@ bool gt_is_run()
     return rf::multi_get_game_type() == rf::NetGameType::NG_TYPE_RUN;
 }
 
-bool gt_is_bm()
+bool gt_is_bag()
 {
-    return rf::multi_get_game_type() == rf::NetGameType::NG_TYPE_BM;
+    return rf::multi_get_game_type() == rf::NetGameType::NG_TYPE_BAG;
 }
 
-bool gt_is_tbm()
+bool gt_is_tbag()
 {
-    return rf::multi_get_game_type() == rf::NetGameType::NG_TYPE_TBM;
+    return rf::multi_get_game_type() == rf::NetGameType::NG_TYPE_TBAG;
 }
 
 bool gt_is_bagman_any()
 {
-    return gt_is_bm() || gt_is_tbm();
+    return gt_is_bag() || gt_is_tbag();
 }
 
 static HillInfo* esc_find_hill_by_role(HillRole role)
@@ -2016,7 +2016,7 @@ CodeInjection send_team_score_patch{
             regs.ax = blue_score;
             regs.eip = 0x00472176; // use stock game packet send
         }
-        else if (gt_is_tbm()) {
+        else if (gt_is_tbag()) {
             const uint16_t red_score = (uint16_t)std::clamp(bagman_get_red_team_score(), 0, 0xFFFF);
             const uint16_t blue_score = (uint16_t)std::clamp(bagman_get_blue_team_score(), 0, 0xFFFF);
             regs.si = red_score;
@@ -2037,7 +2037,7 @@ CodeInjection process_team_score_patch{
             multi_koth_set_red_team_score(red_score);
             multi_koth_set_blue_team_score(blue_score);
         }
-        else if (gt_is_tbm()) {
+        else if (gt_is_tbag()) {
             int red_score = regs.esi;
             int blue_score = regs.edi;
             bagman_set_red_team_score(red_score);