+++ game_patch/rf/multi.h
@@ -141,8 +141,8 @@ namespace rf
         NG_TYPE_REV = 5,    // Revolt, as of AF v1.2
         NG_TYPE_RUN = 6,    // Run, as of AF v1.2
         NG_TYPE_ESC = 7,    // Escalation, as of AF v1.2
-        NG_TYPE_BM = 8,     // Bagman, as of AF v1.4
-        NG_TYPE_TBM = 9,    // Team Bagman, as of AF v1.4
+        NG_TYPE_BAG = 8,    // Bagman, as of AF v1.4
+        NG_TYPE_TBAG = 9,   // Team Bagman, as of AF v1.4
         // Sentinel: UNK must always be the last entry.
         NG_TYPE_UNK
     };